Moved `check_configs` function to `utils/train_utils.py` for better modularity
